PS-462, PS-398 - Add vested balance query, and expose via endpoint

* PS-462, PS-398 - Add vested balance query, and expose via endpoint

* Fix using statements

* Merged develop into feature/PS-462-BalanceTotal

+++ src/services/src/Demoulas.ProfitSharing.Common/Contracts/Services/ParticipantTotalRatioDto.cs
@@ -8,5 +8,5 @@ namespace Demoulas.ProfitSharing.Common.Contracts.Services;
 public class ParticipantTotalRatioDto
 {
     public required long Ssn { get; set; }
-    public required double Ratio { get; set; }
+    public required Decimal Ratio { get; set; }
 }
